Ensure child fragments consider parent menu visibility
Setting the menu visibility of a parent fragment to false should cause
children fragment menu visibility to be false. If the parent visibility
is set back to true, the child fragment should return to its previous
state. This also means that setting the menu visibility of the parent to
false will cause the child to get no callback from the
onCreateOptionsMenu() method.

Test: Added MenuVisibilityFragmentTest and OptionsMenuFragmentTest
Bug: 153593580
Change-Id: Icc41610f7a7dd4f7f181806035e6c6ca96aa2dba