Fix bug in generateReleaseNotes.py where new libraries found no commits
In the case of a new library, GitClient.getGetLog made the incorrect
assumption that n would be populated with an n > 0.  However,
this was actually not the case for new libraries, so the
commit log with run with `-n 0` and return no commits.

This changes that behavior such that n=None is equivalent
to n=infinity, otherwise, it will search with the value of
n that was passed it.

If you run the test, you will see that it stops
complaining about not finding commits for ui-rxjava2 (for example)

Test: ./generateReleaseNotes.py 1586970000000
Test: ./testReleasenotes.py
Change-Id: I9a209b2126d1683a334eecedd371978ad5c1b44d