Introduce RemoveUnusedColumns annotation
This annotation can be used when a Cursor mismatch happens where
the query returns more columns than the ones that will be used in
the response object.

This is a convenience for developers who can still write * projection
for future proofing while not paying the cost of pulling unused
columns from the database.

It can be used along with a @Query, @Dao or @Database annotation and
will scope the functionality similar to the TypeConverters.

When enabled, query is wrapped in a
SELECT <usedColumns> FROM (<originalQuery>).

SQLite automatically flattens this query so it does not become a perf
problem in the SQLite side either.

Bug: 153387066
Test: QueryMethodProcessorTest, RemoveUnusedColumnsTest

Change-Id: Ibdb7637b1731de155b7fec2783020e8857936d4c