Use under and over fragments for relative positioning
The index that a newly added fragment's view should
be added to its container should be relative to
*all* other fragments in that container. This means
it should be above any fragments that are before it
in mAdded order and below any fragments that are
after it in mAdded order.

If no other fragment exists in the container, the
fragment's view should be added at the end of the
container (i.e., an index of -1) so that the fragment
is after any non-fragment views in the same container.

Test: updated FragmentStoreTest tests
BUG: 147749580
BUG: 154241885
Change-Id: I51f5e4c5d30d2ead62c27b2d44c68d704f31713e