Extract RTL spinner test into its own class
The previous implementation that was setting locale and waiting
for activity to be recreated didn't work on v26 and v27.

This one has a separate activity that is configured in
attachBaseContext() to use an RTL locale.A couple of helper methods in
the existing AppCompatSpinnerTest have been made static so that they
can be used in the new test.

Bug: 154294002
Test: ./gradlew :appcompat:appcompat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.appcompat.widget.AppCompatSpinnerTest
Test: ./gradlew :appcompat:appcompat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.appcompat.widget.AppCompatSpinnerRtlTest
Change-Id: I382592a0acfe4b47c9740ff52aeb9fa4e5fc96f6