Add partial gesture InputDispatcher API
This allows developers of gestures to send individual touch events to
an InputDispatcher, potentially splitting the gesture over multiple
GestureScope method invocations. The individual motion events are
combined into one gesture by means of a GestureToken, which basically
just keeps track of the time of the down event, and the time and
position of the last event.

Bug: 152477560
Test: Added new tests for AndroidInputDispatcher.
Relnote: "Adds `sendDown`, `sendMove`, `sendUp` and `sendCancel` methods
to `InputDispatcher`, and `GestureToken` to combine these individual
touch events to one gesture."

Change-Id: Ibd278f19344d41a1302bae93f3d72fb1e8b28c2e