Add accessibility support for custom actions
We use a predefined 32 resource ids for custom actions, as in Viewcompat
accessibility action api. If this virtual node had custom action
id assignment before, we try to keep the id unchanged for the same
action (identified by action label). This way, we can minimize the
influence of custom action change between custom actions are presented
to the user and actually performed.

Test: tested with the Toggleable widget. The fake custom action (click
to change state) appears in Talkback local context menu and does the job.

Change-Id: I594eb6b13d53747402fe6c53d031eb309b44ac61