Annotations module for contentaccess.
The api design is not final, though it does match out current plan.

The docs aren't written yet because this isn't finalized, although
param descriptions are there.

Notice there is no onconflict strategy for insertions this is because
content resolver do not support them out of the box, we may add a
wrapper that queries for existence first etc... but that's a design
decision to be made in the future.

Test: ran the "build" task.
Change-Id: If4f05afdbd8a68d9e73ffbe21f07baafb6491a32