Fix CoreText onPositioned callback performance
Before change aosp/1286377 we were calculating the global coordinates every time onPositioned callback called for the text, except for the first execution. After aosp/1286377 we started doing it every time which causes regression in a lot of first_layout benchmarks. But in fact we only need to calculate the global coordinates when there is a currently active selection, I added such check and now this should improve the performance for both first_layout and layout benchmarks.

Bug: 154577232
Test: manually to verify the scrolling the content with an active selection still moves selection handles as well, verified few benchmarks
Change-Id: Ib0fc7a579ec38541ae5a75cbbe2f6d63139babd6