Temporarily remove Table
The CL temporarily removes the Table layout, until we have resources to
prioritize adding it back with a refreshed API. The current API has fell
behind and remained outdated after the more recent changes to ui-layout
APIs. This makes the component have a high maintainance cost - for
example it is currently blocking the removal of the deprecated DrawNode
and DataNode.

Test: n/a
Bug: 150139220
Relnote: "Table layout was removed temporarily until we will make it available again with a refreshed API."
Change-Id: Id88a7d9c0a4de2c1abd4f030c27f77f73fa21bb3