Annotate BuildCompat version check methods as such
This CL updates the various BuildCompat version
check methods with the new annotation in androidx
for this purpose, which will let lint stop hardcoding
knowledge about these methods (as well as custom
checker method added by users in their own code,
the main use case for this).

Note that this also switches core to depend on the
latest annotation library.

Test: In lint
Bug: 120255046

Change-Id: I300aa3b7bb241e603be3c9e2f5b8f621bc516ba2