Restore ScrollerPosition using saved instance state
We just save raw scroller position in pixels. This solution is not ideal as if items have different sizes in portrait and landscape(for example text with wrap content) the restored position could be not pointing on the same item it was before. But it could be good enough for Scroller component. In AdapterList we will have more information about the children and their sizes so we would be able to restore the scroller position smarter there.

Relnote: HorizontalScroller and VerticalScroller not restores their scroll position using saved instance state.
Bug: 155075048
Test: manually, new test
Change-Id: Ia0fae1fe0df60d85302255acb8577cd8e769d297