Invoke Composables in PreviewActivity only in debug mode
PreviewActivity sets its content to a @Composable passed via intent, and
that could potentially be exploited in released apps that depend on
ui-tooling to use @Preview capabilities. PreviewActivity could be used, for
instance, to execute @Composable functions of features sill hidden under a
flag.

For that reason, we should make the PreviewActivity no-op if the APK is not
a debug build.

Bug: 155017180
Test: will be done on the Android Studio side.
Change-Id: Ie45a2bb8d0266f0522973127502b9524f6a8e65d