Semantics: remove 'container = true'/isSemanticBoundary.
In the new design there will be a single explicit merging primitive
'accessibilityGroup' -- a first step to get there is to delete the
concept of 'container'.  After this deletion, every layout node with
semantics on it behaves as if it has 'container = true' (the old
behavior where 'container = false' nodes are auto-merged-up to the
nearest container no longer exists).

This improves accessibility by making static text nodes no longer merged
up to the root, which was a request from accessibility-team.  It also
means tests can be written a bit more simply.

Notes:
- I needed to fix up some existing tests that were relying on the
  testTag getting merged up to the root semantics node: by either adding
  a Box or when that didn't work, by switching them to a new findRoot
  finder.

- For now, I left the 'container' parameter as a no-op on the
  legacy Semantics() single-child composable (along with its 90 uses
  mostly in tests).  I plan to batch the labor of deleting them along
  with changing tests to use modifiers in a later patch.

Fixes: 157108946
Bug: 153198389
Test: existing tests
Change-Id: I7194a4bc6761a14fff1444481f0652e08b54eca2