UseCaseMediator to LifecycleCamera
Refactor UseCaseMediator so that it becomes the LifecycleCamera class.
This way there is the clear distinction of the LifecycleCamera which is
responsible for controlling the camera start/stop via a Lifecycle.

This is part of the effort to remove the logic from the CameraX class
that doesn't need to be a global.

For more context see: go/camerax-camera-hierarchy

Bug: b/141349818
Test: android.camera.* tests
      manual test of camera-core integration app

Change-Id: I312ab6d2d60d0c5ccceb6142d7e68f115de61c23