Implement focus callbacks on AndroidComposeView
We had only implemented dispatchWindowFocusChanged. This CL adds an implementation for onFocusChanged. This CL also fixes a bug that incorrectly changed the root focus state whenever the window lost focus. We now retain the focus state when the window loses focus, and only clear the focus state if the view loses focus.

Bug: 151335411
Fixes: 151335411
Fixes: 157655888
Test: Ran the app locally to verify that this fixed the bug, and added new tests.
Change-Id: I057167583d2ea499721523ff9a5a4fd7f48f5eb2