Refactored drawRoundRect to consume single Radius parameter
Relnote: "Refactored Radius class to be an
inline class. Removed companion creation
methods in favor of function constructor
with default parameter to have the radius
along the y-axis match that of the mandatory
x-axis radius parameter.

Updated DrawScope.drawRoundRect to consume
a single Radius parameter instead of 2 separate
float values for the radius along the x and y
axis"

Change-Id: I46d1b6c89a0f738304c915ce7ee52b621e10302f
Fixes: 157661009
Test: re-ran compose tests