Fail CameraControl#startFocusAndMetering if no valid meteringRectangles
Currently,  CameraX will continue the focus and metering action
even when MeteringRectangles are all invalid (width or height is 0)
Although it does not cause issue so far as the HAL seems simply
ignore it.  But it does potentially error-prone.  We should
return a failed future immediately.

Relnote: "CameraControl#startFocusAndMetering will fail if none of the
specified MeteringPoint can generate valid metering rectangles."
Bug: NA
Test: FocusMeteringControlTest
Change-Id: Id53ce08eb463ae641765eb91bacb5b466ba6a008