Ensure child views are created before restoring view state
We recently made to a change to ensure that parents have their view
state restored before their children (aosp/1317939). This had the side
effect of restoring the parent view before the child view had been
created. Attempting to access the child view in the onViewRestored of
the parent would result in an error.

This change adds an internal VIEW_CREATED state that is between CREATED
and ACTIVITY_CREATED. This allows us to create the child views right
after the parent view, and ensure they are available when the parent
view state is restored.

Test: changed NestedFragmentRestoreTest, no new failing tests with
USE_STATE_MANAGER
Bug: 158582077

Change-Id: Ie1c9c7c1e9b07de502ef67401c8c118ec09e9c26