Changes to content access + TODO's done:
- query parameter is now called projection, and is a list instead
of string, giving the user the option of selecting a subset of the
fields available in the POJO.
- Specifying fields in projections and orderBy as well as comparisons
in selection are now done by referencing the actual column names as
opposed to the kotlin/java field names in the entity.
- Users can annotate fields in POJOs with @ContentColumn to specify
which column they are referencing, otherwise defaults to the column field
name.
- Added helper error functions
- Error propagation now happens through an ErrorIndicator object instead
of a Pair<Type, Boolean>.
- Added some error checking mentioned in some TODOs.

Test: Play around with the test app.
Change-Id: Ibccf29cbbc9ec45c3a49eb46e2250ac95c02d2bb