Fix Android Studio can not be updated issue.
Updating Android Studio require several long-running steps,
if users exist in any step, Android Studio can not be updated
any more.

Reason: We only check if the Android Studio install path exists or not,
but creating install path is the first step of updating, which means
it can not guarantee other steps is successful.

Solution: We create `STUDIO_UPDATED` file in install dir after all steps
are successful, and check by this file to see if Android Studio needs to be updated.
Every time we update Android Studio, we need to clean up all installations
to avoid `file already exists` issue.

Since we add `INSTALL_SUCCESSFUL` to indicate AS is installed
successfully, to avoid current users downloading the same AS
because lack of the file, we need to check if license accepted
file exists or not.

Bug: 158862268
Test: ./studiow
Change-Id: I8f0a959093b9f6e794346b0d5e0588ebcb8d8df1