Add relative offsetting of positions
In GestureScope and PartialGestureScope, you can now create Offset
instances with values relative to the component's size. This can be
useful to, say, start a swipe 10% inwards from the top left corner of
the component.

I changed the implementation of `centerX` and `centerY`, such that they
no longer divide the `right` and `bottom` by 2, but the `width` and the
`height`. This is much more intuitive, and only requires explaining that
`right != width` and `bottom != height`.

Fixes: 158291007
Test: ./gradlew ui:ui-test:cC
Relnote: "Added `relativeOffset(x, y)` to `GestureScope` and
`PartialGestureScope`, which produces an Offset instance with values
relative to the component's size. `relativeOffset(.1f, .2f)` means an
Offset where x is 10% of the component's width and y is 20% of the
component's height."

Change-Id: Ia4bf75eeeb252ff3eeec0ee35f28cbae618b1abb