Defensive clean up of AndroidOwnerRegistry after tests
Might be a source of leaks otherwise. It might not be related to the
referred bug, but while investigating that bug I noticed an unusual
amount of WIN DEATHs of Windows at the end of a crashing test. I'm not
sure what exactly that means, but one hypothesis is that the
AndroidOwnerRegistry holds on to AndroidOwners (although that seems
unlikely, given that they're in a WeakHashMap). Anyway, it seems like a
good idea to clean up after a test.

Bug: 159111612
Test: ./gradlew ui:ui-test:cC
Change-Id: If3053b00cc908e71879f92cd7996322d17fdec8a