Remove LifecycleObserver when clearing ResultListener
When calling clearFragmentResultListener, we remove the listener from
the fragmentManager, but we never remove the observer that was added to
the lifecycle. This means that the observer remains active until the
lifecycle is DESTROYED. The observer should be cleared out as well.

Test: Added new test in FragmentResultTest
Bug: 159274993
RelNote: "Calling clearFragmentResultListener now properly clears the
lifecycle observer."

Change-Id: I76e67447d1b858dc5440f51eb097e18c267a0baa