Add ui-animation-tooling library
Create a tooling library to support sharing objects between ui-tooling and
Android Studio. This lightweight library contains the ComposeAnimation
interface, and Compose animations and related data can be parsed into it.
Android Studio gets these objects via reflection or bytecode manipulation,
and the plan is to make it import the ui-animation-tooling prebuilt JAR, so
it will be able to parse the data into ComposeAnimation objects, which makes
them much better to manipulate.

We set the generateDocs property to false, as the library is meant to be
used internally only.

Bug: 159418525
Test: N/A
Change-Id: I3453c77992e7fbf0506a01eab02202cc9b92aefa