Relaxed restriction of size to be greater than zero
Previously if a composable was given a
size of zero, it would cause the
DrawScope instance to throw an exception
as the size was being set to zero.

Having a size of zero is useful in transient
scenarios where you would be animating the
drawing bounds starting from zero to the
final size. By relaxing this restriction
these types of animations are possible

Change-Id: I249b373d18fa701bf74058cc11a2b771569ce268
Fixes: 159721920
Test: Added tests to DrawScope