MediaRouter: Release session for disabled group.
MediaRouter unselected the selected route when it becomes disabled.
So media route providers can notify clients of disconnected routes by
setting disable in route info.

With this CL, the behavior is consistent for the output switcher.
(setting disabled will notify release of session to the output switcher)

Bug: 159646626
Test: Manually by calling setEnabled(false) in
notifyDynamicRoutesChanged. w/ and w/o the output switcher.

Change-Id: Id6574e9ce8e00a2376f801b744ee54667f9a9fa9