Revert "Fix Slice not pinned bug when using SliceCallback"
This reverts commit fa5ff3f38794287b8a8f1f3b85e606138c84216a.

Reason for revert: KeyguardSliceView may crash in its refresh() which binds slice directly.
Since KeyguardSliceView.setupUri() removes LiveData observer and then
re-observes, which calls unpinSlice() and pinSlice() in order. But
in the CL, it postpones unpinSlice() to AsyncTask which may be executed
after pinSlice(). When refresh() is called, the slice might already be
unpinned.

Bug: 159943588
Change-Id: I6f2a48aa2aa430a853b4c24e696cee6061455bfa