Support7Demos: Fix the start position of media for member route
The start position of media when a member route is added is incorrect.
Even though a member route is added after N seconds of video playback,
the video for added route always starts from the beginning.

It is due to that the current position of media is not updated when
the member route is added. This CL fixes the problem.

Test: Locally, and checked the logs
Change-Id: I7d5f2fb4479bd66bb3a01c432adad74eeb840747