Change Default ContentScale for Image
Relnote: "Changed the default ContentScale
parameter for the Image composable
from Inside to Fit. This was done in order
to get behavior to scale up the underlying
Painter if the layout size is larger than
the intrinsic size of the painter while
maintaining the aspect ratio. This
behavior better matches expectations
for providing fixed sizes to the Image
while not affecting the default behavior
if only the intrinsic size is used to
compute the size of the composable."

Fixes: 159838006
Test: Updated ImageTest to optionally provide
ContentScale.Inside as well as removing
explicit usages of ContentScale.Fit

Change-Id: I40ae3af9b6a2efc3d730ea0ba5f457a788eca1f7